requirements.txt
Babel==2.5.1
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.6
django-phonenumber-field==1.3.0
django-rest-framework==0.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.2
phonenumberslite==8.8.4
Pillow==4.3.0
pytz==2017.2
requests==2.18.4
simplejson==3.11.1
uritemplate==3.0.0
urllib3==1.22
django-cors-headers==2.1.0
sorl-thumbnail==12.4.1
sorl-thumbnail-serializer-field==0.2.1
